+++ client/components/SearchCustomer/index.tsx
@@ -14,12 +14,16 @@ export const SearchCustomer = (props: ISearchCustomerProps) => {
 
   const searchData: ISuggestionItem<Customer>[] = useMemo(
     () =>
-      (data?.customers || []).map((customer) => ({
-        key: customer.key,
-        displayValue: customer.name,
-        searchValue: [customer.key, customer.name].join(' '),
-        data: customer
-      })),
+      (data?.customers || []).map(
+        (customer) =>
+          ({
+            key: customer.key,
+            text: customer.name,
+            searchValue: [customer.key, customer.name].join(' '),
+            data: customer,
+            iconName: customer.icon || 'Page'
+          } as ISuggestionItem<Customer>)
+      ),
     [data]
   )
 
@@ -29,6 +33,12 @@ export const SearchCustomer = (props: ISearchCustomerProps) => {
         {...props}
         disabled={loading}
         items={searchData}
+        itemIcons={{
+          style: {
+            marginTop: 8,
+            fontSize: 14
+          }
+        }}
         width={550}
         placeholder={t('common.searchPlaceholder')}
         onClear={() => props.onSelected(null)}