+++ client/pages/Admin/Users/UserForm/index.tsx
@@ -54,7 +54,7 @@ export const UserForm: FunctionComponent<IUserFormProps> = (props: IUserFormProp
             placeholder={t('common.searchPlaceholder')}
             items={adUsers.map((u) => ({
               key: u.id,
-              displayValue: u.displayName,
+              text: u.displayName,
               searchValue: u.displayName,
               data: u
             }))}